requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.0.2
django-allauth==0.41.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.11.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
pkg-resources==0.0.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.1
requests==2.22.0
requests-oauthlib==1.3.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.7
